@@ -20,7 +20,7 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.2.0+3
 
 environment:
-  sdk: ">=2.18.6 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -41,12 +41,12 @@ dependencies:
   url_launcher: ^6.1.7
   flutter_barcode_scanner: ^2.0.0
   font_awesome_flutter: ^10.3.0
-  package_info_plus: ^3.0.2
+  package_info_plus: ^8.1.0
   app_review: ^2.1.1+1
   shared_preferences: ^2.0.15
-  flutter_hooks: ^0.18.5+1
-  share_plus: ^7.0.2
-  image_gallery_saver: ^1.5.0
+  flutter_hooks: ^0.20.5
+  share_plus: ^10.1.1
+  image_gallery_saver: ^2.0.3
   qr_flutter: ^4.0.0
 
 dev_dependencies:
